--- examples/email/src/components/select/index.tsx
@@ -1,1 +1,0 @@
-export { Select, type SelectProps } from './Select';
