--- examples/email/src/constants/index.ts
@@ -1,6 +1,0 @@
-enum EMAIL_BOX {
-  INBOX = "inbox",
-  SENT = "sent",
-}
-
-export {EMAIL_BOX};
