+++ examples/email/src/modules/landingPage/components/LandingPageBanner.tsx
@@ -1,0 +1,5 @@
+const LandingPageBanner = ({ height }: { height: string }) => {
+  return <img src={'/EmailBanner.png'} style={{ height }} />;
+};
+
+export { LandingPageBanner };
