+++ examples/email/src/modules/landingPage/components/SimulateTxText.tsx
@@ -1,0 +1,11 @@
+const SimulateTxText = ({
+  height,
+  width,
+}: {
+  height: string;
+  width: string;
+}) => {
+  return <img src={'/EmailLogo.png'} style={{ height, width }} />;
+};
+
+export { SimulateTxText };
