+++ examples/email/src/pages/EmailPage.tsx
@@ -1,0 +1,7 @@
+import EmailScreen from '@/modules/emailPage/EmailScreen';
+
+const LandingPage = () => {
+  return <EmailScreen />;
+};
+
+export default LandingPage;
