+++ examples/email/src/pages/LandingPage.tsx
@@ -1,0 +1,12 @@
+import { LandingPageScreen } from '../modules/landingPage/LandingPageScreen';
+import { ContentLayout } from '../common/components/ContentLayout';
+
+const LandingPage = () => {
+  return (
+    <ContentLayout>
+      <LandingPageScreen />
+    </ContentLayout>
+  );
+};
+
+export default LandingPage;
